Checked application flow and added scrollview in the login fragment

+++ AndroidDemo/Zebra/sample/src/main/java/zebra/beans/ReviewItem.java
@@ -6,11 +6,11 @@ package zebra.beans;
 public class ReviewItem {
     public int profileImage;
     public String memberId;
-    public float ratingBar;
+    public double ratingBar;
     public String reviewText;
 
 
-    public ReviewItem(int profileImage, String memberId, float ratingBar, String reviewText){
+    public ReviewItem(int profileImage, String memberId, double ratingBar, String reviewText){
         this.profileImage = profileImage;
         this.memberId = memberId;
         this.ratingBar = ratingBar;
